Move Noop Module Test Helpers to top level entry points
This module has shared state. It needs to be external from builds.

This lets us test the built versions of the Noop renderer.
